requirements.txt
langchain
langchain-community
Twilio
Python-dotenv
Rpip equests